requirements.txt
python-docx
num2words
python-telegram-bot